Drying up Controller test.
